Move types to state to be reused by component
