+++ client/src/routes/Session/state/sessionNotificationsState.ts
@@ -1,5 +1,14 @@
 import create from 'zustand';
-import {NotificationProps} from '../components/Notifications/Notification';
+import {IconType} from '../../../common/components/Icons';
+
+export type NotificationProps = {
+  text: string;
+  letter?: string;
+  Icon?: IconType;
+  image?: string;
+  timeVisible?: number;
+  visible?: boolean;
+};
 
 type State = {
   notifications: NotificationProps[];
